# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sanjeevani Backend",
    "description": "Sanjeevani supply-chain disruption system backend.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/vendors": {
      "get": {
        "tags": [
          "vendors"
        ],
        "summary": "List Vendors",
        "operationId": "list_vendors_api_v1_vendors_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vendors/{vendor_id}": {
      "get": {
        "tags": [
          "vendors"
        ],
        "summary": "Get Vendor",
        "operationId": "get_vendor_api_v1_vendors__vendor_id__get",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vendor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/vendors": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Vendors",
        "description": "List vendors from the public directory.",
        "operationId": "get_public_vendors_api_v1_public_vendors_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "verified",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Verified"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicVendorList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/vendors/{vendor_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Vendor",
        "description": "Get a specific vendor with full verification details.",
        "operationId": "get_public_vendor_api_v1_public_vendors__vendor_id__get",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicVendor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/vendors/register": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Register Vendor",
        "description": "Register a new vendor. Validates GSTIN offline, checks for duplicates.",
        "operationId": "register_vendor_api_v1_public_vendors_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicVendor": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "lead_time_days": {
            "type": "integer",
            "title": "Lead Time Days"
          },
          "capacity_hint": {
            "type": "string",
            "title": "Capacity Hint"
          },
          "price_band": {
            "type": "string",
            "title": "Price Band"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "reliability": {
            "type": "object",
            "title": "Reliability"
          },
          "verification": {
            "$ref": "#/components/schemas/VendorVerification"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "city",
          "state",
          "lead_time_days",
          "capacity_hint",
          "price_band",
          "languages",
          "verification"
        ],
        "title": "PublicVendor"
      },
      "PublicVendorList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicVendor"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "PublicVendorList"
      },
      "RegistrationResponse": {
        "properties": {
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "created": {
            "type": "boolean",
            "title": "Created"
          },
          "reason_if_duplicate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason If Duplicate"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VendorVerification"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "created"
        ],
        "title": "RegistrationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Vendor": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "gstin": {
            "type": "string",
            "title": "Gstin"
          },
          "udyam_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Udyam Number"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "reliability_score_0_100": {
            "type": "integer",
            "title": "Reliability Score 0 100"
          },
          "on_time_rate": {
            "type": "number",
            "title": "On Time Rate"
          },
          "orders_completed": {
            "type": "integer",
            "title": "Orders Completed"
          },
          "disputes": {
            "type": "integer",
            "title": "Disputes"
          },
          "dues_paise": {
            "type": "integer",
            "title": "Dues Paise"
          },
          "dues_display": {
            "type": "string",
            "title": "Dues Display"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "category",
          "gstin",
          "phone",
          "city",
          "state",
          "lat",
          "lng",
          "reliability_score_0_100",
          "on_time_rate",
          "orders_completed",
          "disputes",
          "dues_paise",
          "dues_display"
        ],
        "title": "Vendor"
      },
...
      "VendorList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Vendor"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "VendorList"
      },
...
      "VendorRegistration": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gstin": {
            "type": "string",
            "title": "Gstin"
          },
          "udyam_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Udyam Number"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "pincode": {
            "type": "string",
            "title": "Pincode"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "default": ""
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "capacity_hint": {
            "type": "string",
            "title": "Capacity Hint",
            "default": "Medium"
          },
          "capacity_units_per_month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Units Per Month"
          },
          "lead_time_days": {
            "type": "integer",
            "title": "Lead Time Days"
          },
          "price_band": {
            "type": "string",
            "title": "Price Band",
            "default": "Standard"
          }
        },
        "type": "object",
        "required": [
          "name",
          "gstin",
          "category",
          "city",
          "state",
          "pincode",
          "phone",
          "languages",
          "lead_time_days"
        ],
        "title": "VendorRegistration"
      },
      "VendorVerification": {
        "properties": {
          "overall_status": {
            "type": "string",
            "title": "Overall Status"
          },
          "gstin_status": {
            "type": "string",
            "title": "Gstin Status"
          },
          "gstin_masked": {
            "type": "string",
            "title": "Gstin Masked"
          },
          "udyam_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Udyam Status"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/VerificationDetail"
            },
            "type": "array",
            "title": "Details"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "checked_at": {
            "type": "string",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "overall_status",
          "gstin_status",
          "gstin_masked",
          "details",
          "source",
          "checked_at"
        ],
        "title": "VendorVerification"
      },
      "VerificationDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "field",
          "status",
          "description"
        ],
        "title": "VerificationDetail"
      },
...
    }
  }
}